Add unit tests for modular components
