+++ include/locomotion_system.h
@@ -1,6 +1,6 @@
 /**
- * @file hexapod_locomotion_system.h
- * @brief Hexapod Locomotion Control System inspired by OpenSHC
+ * @file locomotion_system.h
+ * @brief Locomotion Control System inspired by OpenSHC
  * @author BlightHunter Team
  * @version 1.0
  * @date 2024
@@ -40,147 +40,22 @@
  * - CONTROL_FREQUENCY: Control frequency (Hz) [Typical: 50-100Hz]
  */
 
-#ifndef HEXAPOD_LOCOMOTION_SYSTEM_H
-#define HEXAPOD_LOCOMOTION_SYSTEM_H
+#ifndef LOCOMOTION_SYSTEM_H
+#define LOCOMOTION_SYSTEM_H
 
 #include <Arduino.h>
 #include <ArduinoEigen.h>
 #include <math.h>
-
-// System configuration
-#define NUM_LEGS 6
-#define DOF_PER_LEG 3
-#define TOTAL_DOF (NUM_LEGS * DOF_PER_LEG)
-
-// Configurable physical parameters
-struct HexapodParameters {
-    // Physical dimensions
-    float hexagon_radius; // Hexagon radius (mm)
-    float coxa_length;    // Coxa length (mm)
-    float femur_length;   // Femur length (mm)
-    float tibia_length;   // Tibia length (mm)
-
-    // Robot characteristics
-    float robot_height;             // Nominal height (mm)
-    float robot_weight;             // Weight (kg)
-    Eigen::Vector3f center_of_mass; // Center of mass [x,y,z] (mm)
-
-    // Angle limits (degrees)
-    float coxa_angle_limits[2];  // [min, max]
-    float femur_angle_limits[2]; // [min, max]
-    float tibia_angle_limits[2]; // [min, max]
-
-    // DH parameters (a, alpha, d, theta_offset)
-    float dh_parameters[NUM_LEGS][DOF_PER_LEG][4];
-
-    // Sensor configuration
-    Eigen::Vector3f imu_calibration_offset; // [roll, pitch, yaw] (grados)
-    float fsr_threshold;                    // FSR contact threshold
-    float fsr_max_pressure;                 // Maximum FSR pressure
-
-    // Control parameters
-    float max_velocity;         // Maximum speed (mm/s)
-    float max_angular_velocity; // Maximum angular speed (degrees/s)
-    float stability_margin;     // Stability margin (mm)
-    float control_frequency;    // Control frequency (Hz)
-
-    // Inverse kinematics parameters
-    struct IKConfig {
-        uint8_t max_iterations = 30;   // OpenSHC default
-        float pos_threshold_mm = 0.5f; //   ”      ”
-        bool use_damping = true;
-        float damping_lambda = 30.0f; // λ ≈ 30mm  (see paper)
-        bool clamp_joints = true;     // ik_clamp_joints
-    } ik;
-
-    // Body compensation parameters
-    struct BodyCompConfig {
-        bool enable = true;         // “IMU body compensation”
-        float kp = 0.6f;            // same order as OpenSHC demos
-        float lp_alpha = 0.10f;     // 1-pole low-pass
-        float max_tilt_deg = 12.0f; // disable beyond this (OpenSHC default)
-    } body_comp;
-};
-
-// Gait type enumerations
-enum GaitType {
-    TRIPOD_GAIT,
-    WAVE_GAIT,
-    RIPPLE_GAIT,
-    METACHRONAL_GAIT,
-    ADAPTIVE_GAIT
-};
-
-// Leg states
-enum LegState {
-    STANCE_PHASE,   // Leg on ground
-    SWING_PHASE,    // Leg moving
-    LIFT_PHASE,     // Lifting leg
-    TOUCHDOWN_PHASE // Lowering leg
-};
-
-// 3D point structure
-struct Point3D {
-    float x, y, z;
-    Point3D(float x = 0, float y = 0, float z = 0) : x(x), y(y), z(z) {}
-};
-
-// Joint angles structure
-struct JointAngles {
-    float coxa, femur, tibia;
-    JointAngles(float c = 0, float f = 0, float t = 0) : coxa(c), femur(f), tibia(t) {}
-};
-
-// IMU data structure
-struct IMUData {
-    float roll, pitch, yaw;          // Orientation (degrees)
-    float accel_x, accel_y, accel_z; // Acceleration (m/s²)
-    float gyro_x, gyro_y, gyro_z;    // Angular velocity (degrees/s)
-    bool is_valid;
-};
-
-// FSR data structure
-struct FSRData {
-    float pressure;     // Measured pressure
-    bool in_contact;    // Contact state
-    float contact_time; // Contact time (ms)
-};
-
-// Hardware interfaces (to be implemented by the user)
-class IIMUInterface {
-  public:
-    virtual ~IIMUInterface() = default;
-    virtual bool initialize() = 0;
-    virtual IMUData readIMU() = 0;
-    virtual bool calibrate() = 0;
-    virtual bool isConnected() = 0;
-};
-
-class IFSRInterface {
-  public:
-    virtual ~IFSRInterface() = default;
-    virtual bool initialize() = 0;
-    virtual FSRData readFSR(int leg_index) = 0;
-    virtual bool calibrateFSR(int leg_index) = 0;
-    virtual float getRawReading(int leg_index) = 0;
-};
-
-class IServoInterface {
-  public:
-    virtual ~IServoInterface() = default;
-    virtual bool initialize() = 0;
-    virtual bool setJointAngle(int leg_index, int joint_index, float angle) = 0;
-    virtual float getJointAngle(int leg_index, int joint_index) = 0;
-    virtual bool setJointSpeed(int leg_index, int joint_index, float speed) = 0;
-    virtual bool isJointMoving(int leg_index, int joint_index) = 0;
-    virtual bool enableTorque(int leg_index, int joint_index, bool enable) = 0;
-};
+#include "model.h"
+#include "pose_controller.h"
+#include "walk_controller.h"
+#include "admittance_controller.h"
 
 // Main locomotion system class
-class HexapodLocomotionSystem {
+class LocomotionSystem {
   private:
     // Robot parameters
-    HexapodParameters params;
+    Parameters params;
 
     // Hardware interfaces
     IIMUInterface *imu_interface;
@@ -214,16 +89,21 @@ class HexapodLocomotionSystem {
     // Kinematics matrices
     Eigen::MatrixXf dh_transforms[NUM_LEGS][DOF_PER_LEG];
 
+    RobotModel model;
+    PoseController *pose_ctrl;
+    WalkController *walk_ctrl;
+    AdmittanceController *admittance_ctrl;
+
     Point3D transformWorldToBody(const Point3D &p_world) const;
     bool setLegJointAngles(int leg_index, const JointAngles &q);
     void reprojectStandingFeet();
 
   public:
     // Constructor
-    HexapodLocomotionSystem(const HexapodParameters &params);
+    LocomotionSystem(const Parameters &params);
 
     // Destructor
-    ~HexapodLocomotionSystem();
+    ~LocomotionSystem();
 
     // Initialization
     bool initialize(IIMUInterface *imu, IFSRInterface *fsr, IServoInterface *servo);
@@ -294,7 +174,7 @@ class HexapodLocomotionSystem {
     bool update();
 
     // Getters
-    const HexapodParameters &getParameters() const { return params; }
+    const Parameters &getParameters() const { return params; }
     Eigen::Vector3f getBodyPosition() const { return body_position; }
     Eigen::Vector3f getBodyOrientation() const { return body_orientation; }
     LegState getLegState(int leg_index) const { return leg_states[leg_index]; }
@@ -302,7 +182,7 @@ class HexapodLocomotionSystem {
     Point3D getLegPosition(int leg_index) const { return leg_positions[leg_index]; }
 
     // Setters
-    bool setParameters(const HexapodParameters &new_params);
+    bool setParameters(const Parameters &new_params);
     bool setControlFrequency(float frequency);
     bool setStepParameters(float height, float length);
 
@@ -330,26 +210,6 @@ class HexapodLocomotionSystem {
     void compensateForSlope();
 };
 
-// Utility functions
-namespace HexapodUtils {
-float degreesToRadians(float degrees);
-float radiansToDegrees(float radians);
-float normalizeAngle(float angle);
-Point3D rotatePoint(const Point3D &point, const Eigen::Vector3f &rotation);
-float distance3D(const Point3D &p1, const Point3D &p2);
-bool isPointReachable(const Point3D &point, float max_reach);
+#include "math_utils.h"
 
-// Rotation matrix functions
-Eigen::Matrix3f rotationMatrixX(float angle);
-Eigen::Matrix3f rotationMatrixY(float angle);
-Eigen::Matrix3f rotationMatrixZ(float angle);
-
-// Mathematical calculations
-Eigen::Vector3f quaternionToEuler(const Eigen::Vector4f &quaternion);
-Eigen::Vector4f eulerToQuaternion(const Eigen::Vector3f &euler);
-Eigen::Vector4f quaternionMultiply(const Eigen::Vector4f &q1, const Eigen::Vector4f &q2);
-Eigen::Vector4f quaternionInverse(const Eigen::Vector4f &q);
-
-} // namespace HexapodUtils
-
-#endif // HEXAPOD_LOCOMOTION_SYSTEM_H
+#endif // LOCOMOTION_SYSTEM_H
