+++ src/walk_controller.cpp
@@ -1,0 +1,49 @@
+#include "walk_controller.h"
+
+WalkController::WalkController(RobotModel &m)
+    : model(m), current_gait(TRIPOD_GAIT), gait_phase(0.0f) {}
+
+bool WalkController::setGaitType(GaitType gait) {
+    current_gait = gait;
+    gait_phase = 0.0f;
+    return true;
+}
+
+bool WalkController::planGaitSequence(float, float, float) {
+    return true;
+}
+
+void WalkController::updateGaitPhase(float dt) {
+    gait_phase += dt;
+    if (gait_phase >= 1.0f)
+        gait_phase -= 1.0f;
+}
+
+Point3D WalkController::footTrajectory(int leg_index, float phase, float step_height, float step_length,
+                                       float stance_duration, float swing_duration, float robot_height,
+                                       const float leg_phase_offsets[NUM_LEGS], LegState leg_states[NUM_LEGS],
+                                       IFSRInterface *fsr, IIMUInterface *imu) {
+    float leg_phase = phase + leg_phase_offsets[leg_index];
+    if (leg_phase >= 1.0f)
+        leg_phase -= 1.0f;
+    Point3D trajectory;
+    float base_angle = leg_index * 60.0f;
+    const Parameters &p = model.getParams();
+    float base_x = p.hexagon_radius * cos(math_utils::degreesToRadians(base_angle)) + p.coxa_length;
+    float base_y = p.hexagon_radius * sin(math_utils::degreesToRadians(base_angle));
+
+    if (leg_phase < stance_duration) {
+        leg_states[leg_index] = STANCE_PHASE;
+        float support_progress = leg_phase / stance_duration;
+        trajectory.x = base_x + step_length * (0.5f - support_progress);
+        trajectory.y = base_y;
+        trajectory.z = -robot_height;
+    } else {
+        leg_states[leg_index] = SWING_PHASE;
+        float swing_progress = (leg_phase - stance_duration) / swing_duration;
+        trajectory.x = base_x + step_length * (swing_progress - 0.5f);
+        trajectory.y = base_y;
+        trajectory.z = -robot_height + step_height * sin(M_PI * swing_progress);
+    }
+    return trajectory;
+}
