+++ tests/ArduinoEigen.h
@@ -1,0 +1,4 @@
+#ifndef ARDUINO_EIGEN_H
+#define ARDUINO_EIGEN_H
+#include <Eigen/Dense>
+#endif
